requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
gunicorn>=22.0.0,<23.0.0
sqlalchemy>=2.0.30,<3.0.0
psycopg[binary]>=3.1.19,<4.0.0
pydantic-settings>=2.4.0,<3.0.0
email-validator>=2.2.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
PyJWT>=2.9.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
